+++ dev/src/crossroads.js
@@ -16,6 +16,22 @@
 
         normalizeFn : null,
 
+        normalizeAsArray : function(){
+            this.normalizeFn = function (req, vals) {
+                return [vals.vals_];
+            };
+        },
+
+        normalizeAsObject : function () {
+            this.normalizeFn = function (req, vals) {
+                return [vals];
+            };
+        },
+
+        normalizeAsArguments : function () {
+            this.normalizeFn = null;
+        },
+
         create : function () {
             return new Crossroads();
         },